Extract repeater.ts to fix circular dependencies (as repeater is used more in other models than in repeat.ts)
